pc/Makefile: Fix dependencies
The original version of this did build a proper dependency graph, which was
probably the reason for removing object files in two different ways both
before and after doing the build. As well, the .cpp → .o rule was a bit
mysterious; this new version just uses the default Gnu Make rule for this.

We also use the more typical GNU names for the variables, and do various
other bits of cleanup.